+++ jafaal/sessions/__init__.py
@@ -6,11 +6,11 @@ device detection, timeout validation, and token rotation.
 """
 
 from .crud import (
+    claim_session_for_rotation,
     create_session,
     delete_idle_sessions,
     delete_session,
     delete_sessions_by_family,
-    edit_session,
     get_session_by_id,
     get_session_by_id_not_expired,
     get_session_with_oauth_state,
@@ -31,6 +31,9 @@ from .utils import (
     edit_session_object,
     get_user_agent,
     parse_user_agent,
+    rotate_session,
+    session_absolute_deadline,
+    session_expires_at,
     validate_session_timeout,
 )
 
@@ -41,13 +44,13 @@ __all__ = [
     "UsersSessionsInternal",
     "UsersSessionsModel",
     "UsersSessionsRead",
+    "claim_session_for_rotation",
     "cleanup_idle_sessions",
     "create_session",
     "create_session_object",
     "delete_idle_sessions",
     "delete_session",
     "delete_sessions_by_family",
-    "edit_session",
     "edit_session_object",
     "get_session_by_id",
     "get_session_by_id_not_expired",
@@ -56,5 +59,8 @@ __all__ = [
     "get_user_sessions",
     "mark_tokens_exchanged",
     "parse_user_agent",
+    "rotate_session",
+    "session_absolute_deadline",
+    "session_expires_at",
     "validate_session_timeout",
 ]